# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Trading Journal API
  version: 0.1.0
  description: >-
    Serverless Trading Journal backend (Cognito + API Gateway + Lambda + DynamoDB + S3).
    All successful responses use the envelope { data, meta, error }. error is null on success.
servers:
  - url: https://api.example.com/prod
    description: Example production endpoint (replace with real deployed URL)
  - url: /
    description: Relative (local or stage-injected)
security:
  - CognitoAuthorizer: []
components:
  securitySchemes:
    CognitoAuthorizer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Envelope:
      type: object
      required: [data, meta, error]
      properties:
        data: {}
        meta: { type: ["object", "null"], additionalProperties: true }
        error:
          anyOf:
            - $ref: '#/components/schemas/Error'
            - type: 'null'
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: array, items: { type: object } }
...
    TradeCreateRequest:
      allOf:
        - $ref: '#/components/schemas/TradeCreateCore'
    TradeCreateCore:
      type: object
      required: [symbol, side, quantity, openDate]
      additionalProperties: false
      properties:
        symbol: { type: string, minLength: 1 }
        side: { type: string, enum: [BUY, SELL] }
        quantity: { type: number, minimum: 0.0000001 }
        openDate: { type: string, pattern: '^\\d{4}-\\d{2}-\\d{2}$' }
        closeDate: { type: [string, 'null'], pattern: '^\\d{4}-\\d{2}-\\d{2}$' }
        entryPrice: { type: [number, 'null'], minimum: 0 }
        exitPrice: { type: [number, 'null'], minimum: 0 }
        stopLoss: { type: [number, 'null'], minimum: 0 }
        takeProfit: { type: [number, 'null'], minimum: 0 }
        commission: { type: [number, 'null'], minimum: 0 }
        fees: { type: [number, 'null'], minimum: 0 }
        riskAmount: { type: [number, 'null'], minimum: 0 }
        setupType: { type: [string, 'null'], maxLength: 64 }
        timeframe: { type: [string, 'null'], maxLength: 32 }
        marketCondition: { type: [string, 'null'], maxLength: 64 }
        tradingSession: { type: [string, 'null'], maxLength: 32 }
        tradeGrade: { type: [string, 'null'], enum: [A,B,C,D,F,null] }
        confidence: { type: [number, 'null'], minimum: 0, maximum: 10 }
        setupQuality: { type: [number, 'null'], minimum: 0, maximum: 10 }
        execution: { type: [number, 'null'], minimum: 0, maximum: 10 }
        emotionalState: { type: [string, 'null'], maxLength: 128 }
        psychology:
          type: object
          additionalProperties: false
          properties:
            greed: { type: boolean }
            fear: { type: boolean }
            fomo: { type: boolean }
            revenge: { type: boolean }
            overconfidence: { type: boolean }
            patience: { type: boolean }
        preTradeNotes: { type: [string, 'null'], maxLength: 4000 }
        postTradeNotes: { type: [string, 'null'], maxLength: 4000 }
        mistakes: { type: array, items: { type: string, maxLength: 64 }, maxItems: 50 }
        lessons: { type: array, items: { type: string, maxLength: 64 }, maxItems: 50 }
        newsEvents: { type: array, items: { type: string, maxLength: 128 }, maxItems: 50 }
        economicEvents: { type: array, items: { type: string, maxLength: 128 }, maxItems: 50 }
        status: { type: [string, 'null'], enum: [OPEN, CLOSED, PARTIAL, CANCELLED, null] }
        tags: { type: array, items: { type: string, maxLength: 32 }, maxItems: 50 }
        images:
          type: array
          maxItems: 20
          items:
            type: object
            additionalProperties: false
            properties:
              id: { type: string }
              url: { type: string }
              base64Data: { type: string, pattern: '^data:image/' }
              timeframe: { type: [string, 'null'], maxLength: 32 }
              description: { type: [string, 'null'], maxLength: 256 }
...
paths:
...
  /trades/{tradeId}:
    get:
      summary: Get trade by id
      tags: [Trades]
      parameters:
        - in: path
          name: tradeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
    put:
      summary: Update trade
      tags: [Trades]
      parameters:
        - in: path
          name: tradeId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TradeCreateRequest' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
    delete:
      summary: Delete trade
      tags: [Trades]
      parameters:
        - in: path
          name: tradeId
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
...